requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
litellm>=1.50.0
pydantic>=2.7.0
httpx>=0.27.0
arxiv>=2.1.0
rich>=13.7.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0
sentence-transformers>=3.0.0
torch>=2.0.0
